requirements.txt
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-peewee==3.0.4
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-WTF==1.0.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
peewee==3.14.8
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==6.2.5
pytz==2021.3
six==1.16.0
speaklater==1.3
toml==0.10.2
Werkzeug==2.0.2
wtf-peewee==3.0.2
WTForms==2.3.1
